+++ pkg/models/pg_bouncer_settings.go
@@ -1,0 +1,7 @@
+package models
+
+type PgBouncerSettings struct {
+	Name      *string `json:"name,omitempty"`
+	Operation *string `json:"operation,omitempty"`
+	Value     *string `json:"value,omitempty"`
+}